Move `ForeignKeyAnnotationBuilder` => `ForeignKey::AnnotationBuilder`
