+++ lib/annotate_rb/model_annotator/foreign_key_annotation.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module ForeignKeyAnnotation
+      autoload :AnnotationBuilder, 'annotate_rb/model_annotator/foreign_key_annotation/annotation_builder'
+    end
+  end
+end
